8 comiple uniswap contracts with pragma 0.8.13
There were a number of issues with inheritance that changed from version
0.5.16 to 0.8.13 that needed to be addressed. For the most part the use
of override on public state variables that where overriding interface
functions. Some redundant functions in the pair interface were removed
as they existed in the ERC20 interface. One stack to deep error was
addressed by adding a struct to hold the parameters passed to the
function. closes #8